changed ResetPlayerAvatar argument from u8 to u16

+++ src/field_player_avatar.c
@@ -2033,7 +2033,7 @@ static bool8 Fishing_StartMinigame(struct Task *task)
     return FALSE;
 }
 
-void ResetPlayerAvatar(u8 gfxId)
+void ResetPlayerAvatar(u16 gfxId)
 {
     struct ObjectEvent *playerObjEvent = &gObjectEvents[gPlayerAvatar.objectEventId];
